% Generated by roxygen2 (4.0.0): do not edit by hand
\name{h5.read.spikes}
\alias{h5.read.spikes}
\title{Read in multielectrode data stored in HDF5.}
\usage{
h5.read.spikes(h5file, ids = NULL, time.interval = 1, beg = NULL,
  end = NULL, corr.breaks, keep.meta = FALSE)
}
\arguments{
\item{h5file}{The hdf5 file to read in.}

\item{ids}{Vector containing names of units to keep (or reject).}

\item{time.interval}{rate (in seconds) which firing rate is estimated.}

\item{beg}{start time for recording (in seconds).}

\item{end}{end time for recording (in seconds).}

\item{corr.breaks}{vector describing distance binning for correlation matrix.}

\item{keep.meta}{If TRUE, we keep metadata in the return object.}
}
\value{
An s object.
}
\description{
Read in the MEA data stored in the HDF5 format.  We can reduce the range
of data both in time (by specifying the beg, end params) or reduce which
units are read in using IDS.
}

