% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isi.R
\name{isi.gamma}
\alias{isi.gamma}
\title{Returns log(shape) and log(rate) of gamma distributions fitted to ISI of the
spike train. The spike train is split into chunks and gamma distribution for
each of them is estimated using maximum-likelihood. The result is a mean of
log of the estimated gamma distributions.}
\usage{
isi.gamma(spike.train, chunk.length = 100)
}
\arguments{
\item{spike.train}{vector of spike times}

\item{chunk.length}{length of chunks used for estimation}
}
\value{
list with log(shape) and log(rate) with the means of fitted gamma
    distributions
}
\description{
The rate approximates mean firing rate, shape is a measure of firing
regularity:
- for firing in bursts: log(shape) < 1
- for poisson distributed firing: log(shape) = 1
- for firing with peak of ISI: log(shape) > 1
}
\details{
Calculation of the measure adopted from:
Y. Mochizuki et al., Similarity in neuronal firing regimes across mammalian
species. Jounal of Neuroscience (2016) in press.
}

