% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isi.R
\name{isi.local.variation}
\alias{isi.local.variation}
\title{Returns coefficient of local variation between consecutivie ISIs, computed
with formula:
Lv = 3 / (n - 1) sum_{i=1}^{n-1} (ISI_i - ISI_{i+1}) ^ 2
    / (ISI_i + ISI_{i+1})^2}
\usage{
isi.local.variation(spike.train)
}
\arguments{
\item{spike.train}{vector of spike times}
}
\value{
coefficient of local variation of ISIs or 0 if not enough spikes
}
\description{
Coefficient has value 0 when there is no variation between ISI, and is close
to 1 for irregular ISI.
}
\details{
The measure prraoposod in:
Shinomoto S, Shima K, Tanji J (2003) Differences in spiking patterns among
cortical neurons. Neural Comput 15:2823–2842.
}

