% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isi.R
\name{isi.spearman.rank.corr}
\alias{isi.spearman.rank.corr}
\title{Returns Spearman's rank correlation coefficient of ISI. The spike train is
split into chunks and the result is a mean of coefficients calculated per
each chunk.}
\usage{
isi.spearman.rank.corr(spike.train, chunk.length = 100)
}
\arguments{
\item{spike.train}{vector of spike times}

\item{chunk.length}{length for chunks used for computation}
}
\value{
Spearman's rank correlation coefficient or 0 if not enough spikes
}
\description{
The coefficient measures monotonicity of ISI, and is calculated according
to formula:
rho = n / (n - 1) * sum_{i=1}^{i=n-1} (r_i - mean(r)) * (r_{i+1} - mean(r))
    / sum_{i=1}^{i=n} (r_i - mean(r)^2)
where r_i is a rank order of ISI i, equal ISI ranks are assigned average
rank
}

