% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcd_mea.R
\name{mcd.data.to.array}
\alias{mcd.data.to.array}
\title{Read txt data file with spike times exported from MCD file with MC_DataTool.}
\usage{
mcd.data.to.array(file, beg = NULL, end = NULL, pre.spike.ms = 0,
  post.spike.ms = 0)
}
\arguments{
\item{file}{input file}

\item{beg}{time in seconds from which data should be read in, NULL if should
read from the beginning}

\item{end}{time in seconds until which data should be read in, NULL if 
should read until the end of the file}

\item{pre.spike.ms}{time in ms before a spike for which the input file
has extra lines with voltage readout, 0 if no pre-spike times in the file}

\item{post.spike.ms}{time in ms after a spike for which the input file
has extra lines with voltage readout, 0 if no post-spike times in the file}
}
\description{
Returns the spike trains and the channel names in the imported recording.
Spikes outside of the time window [BEG, END] (if non-null) are ignored.
Blank lines, pre and post-spike voltages will be ignored.
}

