% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcd_mea.R
\name{mcd.read.spikes}
\alias{mcd.read.spikes}
\title{Creates s object from txt data file exported from MCD file with MC_DataTool.}
\usage{
mcd.read.spikes(filename, ids = NULL, time.interval = 1, beg = NULL,
  end = NULL, pre.spike.ms = 0, post.spike.ms = 0)
}
\arguments{
\item{filename}{input file}

\item{ids}{IDS optional vector of cell numbershat should be analysed -- the 
other channels are read in but then ignored.}

\item{time.interval}{time bin used for calculation of the firing rate}

\item{beg}{time in seconds from which data should be read in, NULL if should
read from the beginning}

\item{end}{time in seconds until which data should be read in, NULL if 
should read until the end of the file}

\item{pre.spike.ms}{time in ms before a spike for which the input file
has extra lines with voltage readout.}

\item{post.spike.ms}{time in ms after a spike for which the input file
has extra lines with voltage readout.}
}
\description{
Returns the spike trains and the channel names in the imported recording.
Spikes outside of the time window [BEG, END] (if non-null) are ignored.
Blank lines, pre and post-spike voltages will be ignored.
}

