% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcd_mea.R
\name{ncl.read.spikes}
\alias{ncl.read.spikes}
\title{Creates s object from the NCL dataset input file.}
\usage{
ncl.read.spikes(filename, ids = NULL, time.interval = 1, beg = NULL,
  end = NULL)
}
\description{
Creates s object from the NCL dataset input file.
}
\seealso{
\code{\link{mcd.read.spikes}}
}

