\name{sjemea-internal}
\alias{read.ms.mm.data.format1}
\alias{read.ms.mm.data.format2}
\alias{jay.filter.for.na}
\alias{jay.filter.for.max}
\alias{jay.filter.for.min}
\alias{my.upper}
\title{Internal sjemea functions}
\description{
  Internal sjemea functions.
}
\usage{
read.ms.mm.data.format1(cellname, posfile)
read.ms.mm.data.format2(cellname, posfile)
mm.readpos(posfile)
jay.filter.for.na(x)
jay.filter.for.max(x, max)
jay.filter.for.min(x, min)
my.upper(x,diag)
}
\details{
  These are not to be called by the user.
}
\keyword{internal}
