% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{sjemea-package}
\alias{sjemea}
\alias{sjemea-package}
\title{Read and process multielectrode array data.}
\format{A data frame with 512 observations on the following 3 variables.
\describe{ \item{list("electrode")}{a numeric vector} \item{list("x")}{a
numeric vector} \item{list("y")}{a numeric vector} }}
\source{
Stafford et al. (2009) Neuron.
}
\description{
Multielectrode array (MEA) analysis in R
}
\details{
Collection of tools for reading in and processing MEA data.
...
}
\examples{
data(litke1layout)
## maybe str(litke1layout) ; plot(litke1layout) ...
}
\author{
\email{sje30@cam.ac.uk}
}
\seealso{
\code{\link{http://damtp.cam.ac.uk/user/eglen/waverepo}}
Layout of the MEA provided by Alan Litke, as used in the 2009 Neuron paper.

For each electrode name, the (x,y) location of the electrode (in units of
um) is given.
}
\keyword{datasets}
\keyword{package}

