% Generated by roxygen2 (4.0.0): do not edit by hand
\name{sun.read.spikes}
\alias{sun.read.spikes}
\title{Read in Neuroxplorer data from the Sun et al (2008) papers}
\usage{
sun.read.spikes(filename, ids = NULL, time.interval = 1, beg = NULL,
  end = NULL, min.rate = 0)
}
\arguments{
\item{filename}{Name of the text file to be read in.}

\item{ids}{Optional vector of cell numbers that can be analysed, rather
than analysing all electrodes in the recording.  Warning: Not implemented
in all readers.}

\item{time.interval}{Bin width (in seconds) for estimating firing rate.
Defaults to 1 second.}

\item{beg}{Optional start time.}

\item{end}{Optional end time.}

\item{min.rate}{Optional minimal firing rate for an electrode to be
accepted.}
}
\description{
This is a modified version of jay.read.spikes, as the file formats are
very similar; the main difference being the way channels are
named, and that electrodes are spaced 200um apart.
}

