% Generated by roxygen2 (4.0.0): do not edit by hand
\name{tiling.corr}
\alias{tiling.corr}
\title{Compute the tiling coefficient}
\usage{
tiling.corr(a, b, dt = 0.05, rec.time = NULL)
}
\arguments{
\item{a}{first spike train (vector of spike times)}

\item{b}{second spike train}

\item{dt}{}

\item{rec.time}{2-vector stating beg and end of recording.  If omitted
we take the min(max) spike time as the beg(end).}
}
\value{
Tiling coefficient, which should be between -1 and 1.
}
\description{
The tiling coefficient is computed for two spike trains A and B.  Spike
trains must be sorted (smallest first.)
}
\author{
Catherine Cutts and Stephen Eglen
}

